Shifted AccessibilityService to new location
