+++ app/src/main/java/com/undistract/BlockerService.kt
@@ -5,6 +5,7 @@ import android.content.Intent
 import android.os.IBinder
 import android.util.Log
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import com.undistract.services.AppBlockerAccessibilityService
 
 class BlockerService : Service() {
     companion object {
